Added support for more safe numeric conversions

+++ src/from.rs
@@ -1,3 +1,3 @@
 //! Implementations of the `Into` trait for the `Element` enum.
 
-mod u8;
+mod integer;
